[Feature][UI Next] Delete Conditions Component (#7999)
* delete condition component
* rename file manage route
